Addition of Light/Mini Variant of my Hosts File
I hope this is OK, but it would be neat to have the less aggressive variant of my hosts file available through rethink. It's a little bit less aggressive and useful as a list for smart TVs and similar devices which require extra leniency. I have differentiated between the two lists in blocklistConfig.json as well.