trie: optimized repr for flag-node
More than 40% of the nodes are flag-nodes with 80% or more storing a single
blocklist flag. And hence take up 32 bits regardless of the no of blocklists
they represent ('cause of the way the 16x16 two-level bit-map is setup).

Encode flags in flag-nodes (which in itself is 8 bit long) as-is for
upto 3 different blocklists (24 bits in total). So, for over 80% of flag-nodes
that encode just one blocklist, the size goes down from 32 to 8 bits.

Since the encoding is different (8-bit flags as-is vs 16x16 bit-map), the
decoding of these bits is different too. No special bit differentiates one
encoding from another (there's no way to do so). And so, lengths are used:
That is, if the length is 32 bits or more, then it is decoded as 16x16 bit-map.
If not, it is decoded in 8-bit ranges. This is also a reason why no more than
3 blocklist flags (8 bit in size) are encoded as-is (4th blocklist flag would
result in a length of 32 bits, hence causing ambiguity in how to decode it).