Changes
I have found some lists to be more aggressive than what is stated with the number and some lists to be less aggressive. Unified blocklist for example has been very pleasant to use but small ublock filter has been a nightmare. Light switch tracking extended is a very aggressive list stated by the dev so I think it deserves a 2 rather then a 1.